+++ test/user/userUpdateTest.js
@@ -11,6 +11,9 @@ var TokenModel = mongoose.model('Token');
 var User = require('../../models/user');
 var UserModel = mongoose.model('User');
 
+require('../../models/event');
+var EventModel = mongoose.model('Event');
+
 require('sinon-mongoose');
 
 describe("user update tests", function() {
@@ -513,6 +516,154 @@ describe("user update tests", function() {
         should.exist(user);
         user.should.have.property('id').that.equals(userId.toString());
       })
+      .end(done);
+  });
+
+  it('should add recent event for admin user', function(done) {
+    mockTokenWithPermission('UPDATE_EVENT');
+
+    var mockUser = new UserModel({
+      _id: userId,
+      username: 'test',
+      displayName: 'test',
+      active: true,
+      authentication: {
+        type: 'local'
+      }
+    });
+
+    sandbox.mock(UserModel)
+      .expects('findById')
+      .chain('exec')
+      .yields(null, mockUser);
+
+    var mockEvent = new EventModel({
+      _id: 1,
+      name: 'Mock Event'
+    });
+
+    sandbox.mock(EventModel)
+      .expects('findById')
+      .twice()
+      .onFirstCall()
+      .yields(null, mockEvent)
+      .onSecondCall()
+      .yields(null, mockEvent);
+
+    sandbox.mock(UserModel)
+      .expects('findByIdAndUpdate')
+      .withArgs(userId, {recentEventIds: [1]}, {new: true})
+      .yields(null, mockUser);
+
+    request(app)
+      .post('/api/users/' + userId.toString() + '/events/1/recent' )
+      .set('Accept', 'application/json')
+      .set('Authorization', 'Bearer 12345')
+      .expect(200)
+      .expect('Content-Type', /json/)
+      .expect(function(res) {
+        var user = res.body;
+        should.exist(user);
+        user.should.have.property('id').that.equals(userId.toString());
+      })
+      .end(done);
+  });
+
+  it('should add recent event for acl user', function(done) {
+    mockTokenWithPermission('NO_ADMIN_PERMISSION');
+
+    var mockUser = new UserModel({
+      _id: userId,
+      username: 'test',
+      displayName: 'test',
+      active: true,
+      authentication: {
+        type: 'local'
+      }
+    });
+
+    sandbox.mock(UserModel)
+      .expects('findById')
+      .chain('exec')
+      .yields(null, mockUser);
+
+    var eventAcl = {};
+    eventAcl[userId.toString()] = 'OWNER';
+    var mockEvent = new EventModel({
+      _id: 1,
+      name: 'Mock Event',
+      acl: eventAcl
+    });
+
+    sandbox.mock(EventModel)
+      .expects('findById')
+      .twice()
+      .onFirstCall()
+      .yields(null, mockEvent)
+      .onSecondCall()
+      .yields(null, mockEvent);
+
+    sandbox.mock(UserModel)
+      .expects('findByIdAndUpdate')
+      .withArgs(userId, {recentEventIds: [1]}, {new: true})
+      .yields(null, mockUser);
+
+    request(app)
+      .post('/api/users/' + userId.toString() + '/events/1/recent' )
+      .set('Accept', 'application/json')
+      .set('Authorization', 'Bearer 12345')
+      .expect(200)
+      .expect('Content-Type', /json/)
+      .expect(function(res) {
+        var user = res.body;
+        should.exist(user);
+        user.should.have.property('id').that.equals(userId.toString());
+      })
+      .end(done);
+  });
+
+  it('should fail to add recent event for user not in event', function(done) {
+    mockTokenWithPermission('NO_ADMIN_PERMISSION');
+
+    var mockUser = new UserModel({
+      _id: userId,
+      username: 'test',
+      displayName: 'test',
+      active: true,
+      authentication: {
+        type: 'local'
+      }
+    });
+
+    sandbox.mock(UserModel)
+      .expects('findById')
+      .chain('exec')
+      .yields(null, mockUser);
+
+    var mockEvent = new EventModel({
+      _id: 1,
+      name: 'Mock Event',
+      acl: {}
+    });
+
+    sandbox.mock(EventModel)
+      .expects('findById')
+      .twice()
+      .onFirstCall()
+      .yields(null, mockEvent)
+      .onSecondCall()
+      .yields(null, mockEvent);
+
+    sandbox.mock(UserModel)
+      .expects('findByIdAndUpdate')
+      .withArgs(userId, {recentEventIds: [1]}, {new: true})
+      .yields(null, mockUser);
+
+    request(app)
+      .post('/api/users/' + userId.toString() + '/events/1/recent' )
+      .set('Accept', 'application/json')
+      .set('Authorization', 'Bearer 12345')
+      .expect(403)
       .end(done);
   });